requirements.txt
python-pptx==1.0.2
pywin32>=300; sys_platform == 'win32'
